mobileBackend/requirements.txt
appnope==0.1.3
asgiref==3.5.0
asttokens==2.0.5
autopep8==1.6.0
backcall==0.2.0
backports.zoneinfo==0.2.1
blessings==1.7
bpython==0.22.1
certifi==2021.10.8
charset-normalizer==2.0.12
curtsies==0.3.10
cwcwidth==0.1.6
decorator==5.1.1
Django==4.0.3
django-extensions==3.1.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
executing==0.8.3
greenlet==1.1.2
idna==3.3
install==1.3.5
ipython==8.2.0
jedi==0.18.1
matplotlib-inline==0.1.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.29
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.8.0
Pygments==2.11.2
PyJWT==1.7.1
pytz==2022.1
pyxdg==0.27
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
toml==0.10.2
traitlets==5.1.1
typing_extensions==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
